pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pawn-queue"
version = "0.1.0"
description = "S3-compatible pub/sub queue library with no AWS account required"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Davide Dal Farra" }]
keywords = ["s3", "queue", "pubsub", "messaging", "hetzner", "minio"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: AsyncIO",
    "Topic :: System :: Networking",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "aioboto3>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "moto[s3,server]>=5.0",
    "flask>=3.0",
    "flask-cors>=4.0",
    "black>=24.0",
    "ruff>=0.4",
]
e2e = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-timeout>=2.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pawn_queue*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
# Run e2e tests with: pytest e2e/ -v --timeout=120

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]
